pyproject.toml
[project]
name = "astrologer-mcp"
version = "1.0.0"
description = "MCP server for vedic astrological insights on Puch AI powered by OpenAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.11.2",
    "python-dotenv>=1.1.1",
    "httpx>=0.25.0",
    "openai>=1.3.0",
]
